pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentpolicy"
version = "0.1.0"
description = "Runtime policy enforcement for AI agent sessions"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
dependencies = [
    "agentbudget>=0.2.3",
    "PyYAML>=6.0",
]
authors = [
    {name = "Rohan Dahal"},
]
keywords = ["ai", "agents", "policy", "guardrails", "runtime", "security"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
agentpolicy = "agentpolicy.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["agentpolicy*"]
